flask/requirements.txt
pandas
flask
mysql-connector-python